feat: Organize reminder view controller
ReminderViewController
- Extract data source-related behaviors into a separate file `ReminderViewController+DataSource.swift` as extension.

ReminderListViewController.swift
- Remove lines of code not using.
