+++ JustThree/ViewController/ReminderViewController/ReminderViewController+DataSource.swift
@@ -1,0 +1,68 @@
+//
+//  ReminderViewController+DataSource.swift
+//  JustThree
+//
+//  Created by Sanghun Park on 28.09.22.
+//
+
+import UIKit
+
+extension ReminderViewController {
+    typealias DataSource = UICollectionViewDiffableDataSource<Section, Row>
+    typealias Snapshot = NSDiffableDataSourceSnapshot<Section, Row>
+    
+    private func cellRegistration() -> UICollectionView.CellRegistration<UICollectionViewListCell, Row> {
+        return .init { (cell: UICollectionViewListCell, indexPath: IndexPath, row: Row) in
+            
+            let section = self.section(for: indexPath)
+            switch (section, row) {
+            case (_, .header(let title)):
+                cell.contentConfiguration = self.headerConfiguration(for: cell, with: title)
+            case (.view, _):
+                cell.contentConfiguration = self.defaultConfiguration(for: cell, at: row)
+            case (.title, .editText(let title)):
+                cell.contentConfiguration = self.titleConfiguration(for: cell, with: title)
+            case (.date, .editDate(let date)):
+                cell.contentConfiguration = self.dateConfiguration(for: cell, with: date)
+            case (.notes, .editText(let notes)):
+                cell.contentConfiguration = self.notesConfiguration(for: cell, with: notes)
+            default:
+                fatalError("Unexpected combination of section and row.")
+            }
+            cell.tintColor = .justThreePrimaryTint
+        }
+    }
+    
+    func makeDataSource() -> DataSource {
+        let cellRegistration = cellRegistration()
+        return DataSource(collectionView: collectionView) { (collectionView: UICollectionView, indexPath: IndexPath, itemIdentifier: Row) in
+            return collectionView.dequeueConfiguredReusableCell(using: cellRegistration, for: indexPath, item: itemIdentifier)
+        }
+    }
+    
+    func updateSnapshotForEditing() {
+        var snapshot = Snapshot()
+        snapshot.appendSections([.title, .date, .notes])
+        snapshot.appendItems([.header(Section.title.name), .editText(reminder.title)], toSection: .title)
+        snapshot.appendItems([.header(Section.date.name), .editDate(reminder.dueDate)], toSection: .date)
+        snapshot.appendItems([.header(Section.notes.name), .editText(reminder.notes)], toSection: .notes)
+        dataSource.apply(snapshot)
+    }
+    
+    
+    func updateSnapshotForViewing() {
+        var snapshot = Snapshot()
+        snapshot.appendSections([.view])
+        snapshot.appendItems([.header(""), .viewTitle, .viewDate, .viewTime, .viewNotes], toSection: .view)
+        dataSource.apply(snapshot)
+    }
+    
+    private func section(for indexPath: IndexPath) -> Section {
+        let sectionNumber = isEditing ? indexPath.section + 1 : indexPath.section
+        guard let section = Section(rawValue: sectionNumber) else {
+            fatalError("Unable to find matching section")
+        }
+        return section
+    }
+    
+}
