+++ JustThree/ViewController/ReminderViewController/ReminderViewController.swift
@@ -7,11 +7,7 @@
 
 import UIKit
 
-//private let reuseIdentifier = "Cell"
-
 class ReminderViewController: UICollectionViewController {
-    private typealias DataSource = UICollectionViewDiffableDataSource<Section, Row>
-    private typealias Snapshot = NSDiffableDataSourceSnapshot<Section, Row>
 
     var reminder: Reminder {
         didSet {
@@ -21,7 +17,7 @@ class ReminderViewController: UICollectionViewController {
     var workingReminder: Reminder
     var isAddingNewReminder = false
     var onChange: (Reminder)->Void
-    private lazy var dataSource: DataSource = makeDataSource()
+    lazy var dataSource: DataSource = makeDataSource()
     
     init(reminder: Reminder, onChange: @escaping (Reminder)->Void) {
         self.reminder = reminder
@@ -68,35 +64,6 @@ class ReminderViewController: UICollectionViewController {
         }
     }
     
-    func cellRegistration() -> UICollectionView.CellRegistration<UICollectionViewListCell, Row> {
-        return .init { (cell: UICollectionViewListCell, indexPath: IndexPath, row: Row) in
-            
-            let section = self.section(for: indexPath)
-            switch (section, row) {
-            case (_, .header(let title)):
-                cell.contentConfiguration = self.headerConfiguration(for: cell, with: title)
-            case (.view, _):
-                cell.contentConfiguration = self.defaultConfiguration(for: cell, at: row)
-            case (.title, .editText(let title)):
-                cell.contentConfiguration = self.titleConfiguration(for: cell, with: title)
-            case (.date, .editDate(let date)):
-                cell.contentConfiguration = self.dateConfiguration(for: cell, with: date)
-            case (.notes, .editText(let notes)):
-                cell.contentConfiguration = self.notesConfiguration(for: cell, with: notes)
-            default:
-                fatalError("Unexpected combination of section and row.")
-            }
-            cell.tintColor = .justThreePrimaryTint
-        }
-    }
-    
-    private func makeDataSource() -> DataSource {
-        let cellRegistration = cellRegistration()
-        return DataSource(collectionView: collectionView) { (collectionView: UICollectionView, indexPath: IndexPath, itemIdentifier: Row) in
-            return collectionView.dequeueConfiguredReusableCell(using: cellRegistration, for: indexPath, item: itemIdentifier)
-        }
-    }
-    
     @objc func didCancelEdit() {
         workingReminder = reminder
         setEditing(false, animated: true)
@@ -107,36 +74,12 @@ class ReminderViewController: UICollectionViewController {
         updateSnapshotForEditing()
     }
     
-    private func updateSnapshotForEditing() {
-        var snapshot = Snapshot()
-        snapshot.appendSections([.title, .date, .notes])
-        snapshot.appendItems([.header(Section.title.name), .editText(reminder.title)], toSection: .title)
-        snapshot.appendItems([.header(Section.date.name), .editDate(reminder.dueDate)], toSection: .date)
-        snapshot.appendItems([.header(Section.notes.name), .editText(reminder.notes)], toSection: .notes)
-        dataSource.apply(snapshot)
-    }
-    
     private func prepareForViewing() {
         navigationItem.leftBarButtonItem = nil
         if workingReminder != reminder {
             reminder = workingReminder
         }
         updateSnapshotForViewing()
-    }
-    
-    private func updateSnapshotForViewing() {
-        var snapshot = Snapshot()
-        snapshot.appendSections([.view])
-        snapshot.appendItems([.header(""), .viewTitle, .viewDate, .viewTime, .viewNotes], toSection: .view)
-        dataSource.apply(snapshot)
-    }
-    
-    private func section(for indexPath: IndexPath) -> Section {
-        let sectionNumber = isEditing ? indexPath.section + 1 : indexPath.section
-        guard let section = Section(rawValue: sectionNumber) else {
-            fatalError("Unable to find matching section")
-        }
-        return section
     }
     
 }
